Move jetty-nosql to jetty-mongodb.
Also:

- move NoSqlSessionData to MongoDBSessionData
- change package from org.eclipse.jetty.nosql to org.eclipse.jetty.mongodb
