+++ documentation/jetty-documentation/pom.xml
@@ -205,7 +205,7 @@
     </dependency>
     <dependency>
       <groupId>org.eclipse.jetty.session</groupId>
-      <artifactId>jetty-nosql</artifactId>
+      <artifactId>jetty-mongodb</artifactId>
     </dependency>
     <dependency>
       <groupId>org.eclipse.jetty.websocket</groupId>
