+++ documentation/jetty-documentation/src/main/java/org/eclipse/jetty/docs/programming/server/session/SessionDocs.java
@@ -17,7 +17,7 @@ import java.io.File;
 import java.net.InetSocketAddress;
 
 import org.eclipse.jetty.memcached.session.MemcachedSessionDataMapFactory;
-import org.eclipse.jetty.nosql.mongodb.MongoSessionDataStoreFactory;
+import org.eclipse.jetty.mongodb.MongoSessionDataStoreFactory;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.handler.ContextHandlerCollection;
 import org.eclipse.jetty.server.session.CachingSessionDataStoreFactory;
