+++ javadoc/pom.xml
@@ -258,7 +258,7 @@
     </dependency>
     <dependency>
       <groupId>org.eclipse.jetty.session</groupId>
-      <artifactId>jetty-nosql</artifactId>
+      <artifactId>jetty-mongodb</artifactId>
       <scope>provided</scope>
     </dependency>
     <dependency>
