+++ jetty-bom/pom.xml
@@ -229,7 +229,7 @@
       </dependency>
       <dependency>
         <groupId>org.eclipse.jetty.session</groupId>
-        <artifactId>jetty-nosql</artifactId>
+        <artifactId>jetty-mongodb</artifactId>
         <version>12.0.0-SNAPSHOT</version>
       </dependency>
       <dependency>
