+++ jetty-maven-plugin/pom.xml
@@ -305,7 +305,7 @@
         </exclusion>
         <exclusion>
           <groupId>org.eclipse.jetty.session</groupId>
-          <artifactId>jetty-nosql</artifactId>
+          <artifactId>jetty-mongodb</artifactId>
         </exclusion>
         <exclusion>
           <groupId>org.eclipse.jetty</groupId>
