+++ jetty-session/jetty-mongodb/pom.xml
@@ -5,10 +5,10 @@
     <version>12.0.0-SNAPSHOT</version>
   </parent>
   <modelVersion>4.0.0</modelVersion>
-  <artifactId>jetty-nosql</artifactId>
-  <name>Jetty :: NoSQL Session Managers</name>
+  <artifactId>jetty-mongodb</artifactId>
+  <name>Jetty :: MongoDB Session Manager</name>
   <properties>
-    <bundle-symbolic-name>${project.groupId}.nosql</bundle-symbolic-name>
+    <bundle-symbolic-name>${project.groupId}.mongodb</bundle-symbolic-name>
   </properties>
 
   <dependencies>
@@ -30,21 +30,6 @@
     <dependency>
       <groupId>org.slf4j</groupId>
       <artifactId>slf4j-api</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.jetty</groupId>
-      <artifactId>jetty-slf4j-impl</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.jetty.toolchain</groupId>
-      <artifactId>jetty-test-helper</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.jetty.tests</groupId>
-      <artifactId>test-sessions-common</artifactId>
-      <scope>test</scope>
     </dependency>
   </dependencies>
 </project>
