+++ jetty-session/jetty-mongodb/src/main/config/etc/sessions/mongo/session-store-by-uri.xml
@@ -9,13 +9,12 @@
   <!-- ===================================================================== -->
   <Call name="addBean">
    <Arg>
-    <New id="sessionDataStoreFactory" class="org.eclipse.jetty.nosql.mongodb.MongoSessionDataStoreFactory">      
+    <New id="sessionDataStoreFactory" class="org.eclipse.jetty.mongodb.MongoSessionDataStoreFactory">      
        <Set name="dbName"><Property name="jetty.session.mongo.dbName" default="HttpSessions" /></Set>
        <Set name="collectionName"><Property name="jetty.session.mongo.collectionName" default="jettySessions" /></Set>
        <Set name="gracePeriodSec"><Property name="jetty.session.gracePeriod.seconds" default="3600" /></Set>
        <Set name="savePeriodSec"><Property name="jetty.session.savePeriod.seconds" default="0" /></Set>
-       <Set name="host"><Property name="jetty.session.mongo.host" default="localhost"/></Set>
-       <Set name="port"><Property name="jetty.session.mongo.port" default="27017"/></Set>
+       <Set name="connectionString"><Property name="jetty.session.mongo.connectionString" default="mongodb://localhost"/></Set>
     </New>
    </Arg>
   </Call>
