+++ jetty-session/jetty-mongodb/src/main/java/org/eclipse/jetty/mongodb/MongoSessionDataStoreFactory.java
@@ -11,7 +11,7 @@
 // ========================================================================
 //
 
-package org.eclipse.jetty.nosql.mongodb;
+package org.eclipse.jetty.mongodb;
 
 import java.net.UnknownHostException;
