+++ jetty-session/jetty-mongodb/src/main/java/org/eclipse/jetty/mongodb/package-info.java
@@ -12,7 +12,7 @@
 //
 
 /**
- * Jetty NoSql : Generic Nosql Session Management
+ * Jetty NoSql : MongoDB Integration
  */
-package org.eclipse.jetty.nosql;
+package org.eclipse.jetty.mongodb;
