--- jetty-session/jetty-nosql/src/main/java/org/eclipse/jetty/nosql/mongodb/package-info.java
@@ -1,18 +1,0 @@
-//
-// ========================================================================
-// Copyright (c) 1995-2021 Mort Bay Consulting Pty Ltd and others.
-//
-// This program and the accompanying materials are made available under the
-// terms of the Eclipse Public License v. 2.0 which is available at
-// https://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
-// which is available at https://www.apache.org/licenses/LICENSE-2.0.
-//
-// SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
-// ========================================================================
-//
-
-/**
- * Jetty NoSql : MongoDB Integration
- */
-package org.eclipse.jetty.nosql.mongodb;
-
