+++ tests/test-sessions/test-mongodb-sessions/src/test/java/org/eclipse/jetty/mongodb/ClusteredInvalidateSessionTest.java
@@ -11,16 +11,16 @@
 // ========================================================================
 //
 
-package org.eclipse.jetty.nosql.mongodb;
+package org.eclipse.jetty.mongodb;
 
-import org.eclipse.jetty.server.session.AbstractClusteredSessionScavengingTest;
+import org.eclipse.jetty.server.session.AbstractClusteredInvalidationSessionTest;
 import org.eclipse.jetty.server.session.SessionDataStoreFactory;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
 @Testcontainers(disabledWithoutDocker = true)
-public class ClusteredSessionScavengingTest extends AbstractClusteredSessionScavengingTest
+public class ClusteredInvalidateSessionTest extends AbstractClusteredInvalidationSessionTest
 {
     @BeforeAll
     public static void beforeClass() throws Exception
